bosch/requirements.txt
# Bosch Kaggle Pipeline Requirements
# Most packages are already in Kaggle environment

# Core dependencies (usually pre-installed in Kaggle)
pandas>=1.5.0
numpy>=1.23.0
scikit-learn>=1.2.0
xgboost>=1.7.0

# Additional dependencies
pyyaml>=6.0  # For config file parsing
psutil>=5.9.0  # For memory logging (optional, will work without it)

# Note: If running locally, install with:
# pip install -r requirements.txt

